backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0
stripe==7.8.0
resend==2.17.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.0.0
factory-boy==3.3.0
structlog==23.2.0
slowapi==0.1.9
email-validator==2.1.0
